Replace MySQL's "old_password" with "sha1" for...
...hashing passwords. Make no mistake, both are very insecure today! The
correct way of handling this involves salting, using a more modern
cryptographic hash function and verification on PHP's end instead of
sending in a newly generated hash on each attempt to see if it matches
the value stored in the DB (with a random salt it shouldn't).

*This is a temporary measure, just to get the application in a working
state. It is not recommended for production.*